requirements.txt
# app
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2
django-ordered-field==1.0.0
urllib3==1.26.4


# test
behave==1.2.6
behave-django==1.4.0
pytest-django==4.5.2
pytest-mock==3.8.2
selenium==3.141.0
